app/requirements.txt
requests
pandas
psycopg2-binary
SQLAlchemy
python-dateutil
pytz
APScheduler

